Fix connectToReader / waitRemoval.
Connecting to the reader with a invalid name ("") caused an
infinite loop: If the reader name is empty, we let nfc pick a
reader for us.

waitRemoval: calling disconnect() caused waitRemoval to return
true instantly, even if the card was not removed. Now waitRemoval
now attempt to connect() until it can't, meaning the card was
removed.